requirements.txt
google-genai
pyyaml
python-dotenv
